@@ -70,13 +88,14 @@ setup(
     packages=find_packages("src"),
     package_dir={"": "src"},
     package_data={"": ["src/tfc/py.typed"]},
-    python_requires=">=3.10",
+    python_requires=">=3.11",
     include_package_data=True,
-    ext_modules=[BF],
+    ext_modules=[CMakeExtension("BF")],
+    cmdclass={"build_ext": CMakeBuild},
     install_requires=[
         numpy_version,
-        "jax ~= 0.6.0",
-        "jaxlib ~= 0.6.0",
+        "jax >= 0.6.0",
+        "jaxlib >= 0.6.0",
         "jaxtyping",
         "annotated-types",
         "matplotlib",
